Break the monolithic expression_grammar file up
