+++ recipe/schemas/expression_grammar.py
@@ -1,42 +1,11 @@
-import hashlib
 import re
 from collections import defaultdict
-from datetime import date, datetime
 
-import attr
-import dateparser
 import structlog
-from lark import GrammarError, Lark, Transformer, Tree, Visitor, v_args
-from lark.lexer import Token
-from sqlalchemy import (
-    Boolean,
-    Date,
-    DateTime,
-    Float,
-    Integer,
-    String,
-    and_,
-    between,
-    case,
-    cast,
-    distinct,
-    func,
-    not_,
-    or_,
-    text,
-)
+from sqlalchemy import Boolean, Date, DateTime, Integer, String, text
 from sqlalchemy.ext.declarative.api import DeclarativeMeta
 from sqlalchemy.sql.base import ColumnCollection
-from sqlalchemy.sql.expression import select
 from sqlalchemy.sql.sqltypes import Numeric
-
-from . import engine_support
-from .utils import (
-    calc_date_range,
-    convert_to_end_datetime,
-    convert_to_eod_datetime,
-    convert_to_start_datetime,
-)
 
 SLOG = structlog.get_logger(__name__)
 
@@ -300,953 +269,3 @@ def make_grammar(columns):
     %ignore WS_INLINE
 """
     return grammar
-
-
-@attr.s
-class SQLALchemyValidator(Visitor):
-    """Visit the tree and return descriptive information. Populate
-    a list of errors.
-
-    Args:
-        text (str): A copy of the parsed text for error descriptions
-        forbid_aggregation (bool): Should aggregations be treated as an error
-        drivername (str): The database engine we are running against
-    """
-
-    text = attr.ib()
-    forbid_aggregation = attr.ib()
-    drivername = attr.ib()
-    # Was an aggregation encountered in the tree?
-    found_aggregation = attr.ib(default=False)
-    # What is the datatype of the returned expression
-    last_datatype = attr.ib(default=None)
-    # Errors encountered while visiting the tree
-    errors = attr.ib(factory=list)
-
-    def _data_type(self, tree):
-        # Find the data type for a tree
-        if tree is None:
-            return None
-        if tree.data == "col":
-            dt = self._data_type(tree.children[0])
-        else:
-            dt = tree.data
-        dt = str(dt)  # Convert Tokens to strings
-        if dt == "datetime_end":
-            dt = "datetime"
-        elif dt == "string":
-            dt = "str"
-        elif dt == "boolean":
-            dt = "bool"
-        return dt
-
-    def _add_error(self, message, tree):
-        """Add an error pointing to this location in the parsed string"""
-        tok = None
-        # Find the first token
-        while tree and tree.children:
-            tree = tree.children[0]
-            if isinstance(tree, Token):
-                tok = tree
-                break
-
-        if tok:
-            extra_context = self._get_context_for_token(tok, span=200)
-            message = f"{message}\n\n{extra_context}"
-        self.errors.append(message)
-
-    def _get_context_for_token(self, tok, span=40):
-        pos = tok.start_pos
-        start = max(pos - span, 0)
-        end = pos + span
-        before = self.text[start:pos].rsplit("\n", 1)[-1]
-        after = self.text[pos:end].split("\n", 1)[0]
-        return before + after + "\n" + " " * len(before) + "^\n"
-
-    def _error_math(self, tree, verb):
-        tok1 = tree.children[0].children[0]
-        tok2 = tree.children[1].children[0]
-        self._add_error(f"{tok1.data} and {tok2.data} can not be {verb}", tree)
-
-    def col(self, tree):
-        self.last_datatype = self._data_type(tree)
-
-    def error_add(self, tree):
-        self._error_math(tree, "added together")
-
-    def error_mul(self, tree):
-        self._error_math(tree, "multiplied together")
-
-    def error_sub(self, tree):
-        self._error_math(tree, "subtracted")
-
-    def error_div(self, tree):
-        self._error_math(tree, "divided")
-
-    def error_if_statement(self, tree):
-        args = tree.children
-        # Throw away the "if"
-        args = args[1:]
-
-        # If there's an odd number of args, pop the last one to use as the else
-        if len(args) % 2:
-            else_expr = args.pop()
-        else:
-            else_expr = None
-
-        # The "odd" args should be booleans
-        bool_args = args[::2]
-        # The "even" args should be values of the same type
-        value_args = args[1::2]
-
-        # Check that the boolean args are boolean
-        for arg in bool_args:
-            dt = self._data_type(arg)
-            if dt != "bool":
-                self._add_error("This should be a boolean column or expression", arg)
-
-        # Data types of columns must match
-        value_type = None
-        for arg in value_args + [else_expr]:
-            dt = self._data_type(arg)
-            if dt is not None:
-                if value_type is None:
-                    value_type = dt
-                elif value_type != dt:
-                    self._add_error(
-                        f"The values in this if statement must be the same type, not {value_type} and {dt}",
-                        arg,
-                    )
-
-    def aggr(self, tree):
-        self.found_aggregation = True
-        if self.forbid_aggregation:
-            self._add_error("Aggregations are not allowed in this field.", tree)
-
-    def unknown_col(self, tree):
-        """Column name doesn't exist in the data"""
-        tok1 = tree.children[0]
-        self._add_error(f"{tok1} is not a valid column name", tree)
-
-    def unusable_col(self, tree):
-        """Column name isn't a data type we can handle"""
-        tok1 = tree.children[0]
-        self._add_error(
-            f"{tok1} is a data type that can't be used. Usable data types are strings, numbers, boolean, dates, and datetimes",
-            tree,
-        )
-
-    def error_not_nonboolean(self, tree):
-        """NOT string or NOT num"""
-        self._add_error("NOT requires a boolean value", tree)
-
-    def mixedarray(self, tree):
-        """An array containing a mix of strings and numbers"""
-        self._add_error("An array may not contain both strings and numbers", tree)
-
-    def vector_expr(self, tree):
-        val, comp, arr = tree.children
-        # If the left hand side is a number or string primitive
-        if isinstance(val.children[0], Token) and val.children[0].type in (
-            "NUMBER",
-            "ESCAPED_STRING",
-        ):
-            self._add_error("Must be a column or expression", val)
-
-    def error_aggr(self, tree):
-        """Aggregating a bad data type"""
-        fn = tree.children[0].children[0]
-        dt = self._data_type(tree.children[0].children[1])
-        self._add_error(f"A {dt} can not be aggregated using {fn}.", tree)
-
-    def error_between_expr(self, tree):
-        col, BETWEEN, left, AND, right = tree.children
-        col_type = self._data_type(col)
-        left_type = self._data_type(left)
-        right_type = self._data_type(right)
-        if col_type == "datetime":
-            if left_type == "date":
-                left_type = "datetime"
-            if right_type == "date":
-                right_type = "datetime"
-        if not (col_type == left_type == right_type):
-            self._add_error(
-                f"When using between, the column ({col_type}) and between values ({left_type}, {right_type}) must be the same data type.",
-                tree,
-            )
-
-    def bool_expr(self, tree):
-        """a > b where the types of a and b don't match"""
-        left, _, right = tree.children
-        if isinstance(left, Tree) and isinstance(right, Tree):
-            left_data_type = self._data_type(left)
-            right_data_type = self._data_type(right)
-            if left_data_type == right_data_type == "date":
-                return
-            if left_data_type == right_data_type == "datetime":
-                return
-            if left_data_type in ("date", "datetime") and right_data_type == "string":
-                # Strings will be auto converted
-                return
-            if left_data_type != right_data_type:
-                self._add_error(
-                    f"Can't compare {left_data_type} to {right_data_type}", tree
-                )
-
-    def percentile_aggr(self, tree):
-        """Sum up the things"""
-        percentile, fld = tree.children
-        percentile_val = int(percentile[len("percentile") :])
-        if percentile_val not in (1, 5, 10, 25, 50, 75, 90, 95, 99):
-            self._add_error(
-                f"Percentile values of {percentile_val} are not supported.", tree
-            )
-        if self.drivername == "sqlite":
-            self._add_error("Percentile is not supported on sqlite", tree)
-
-
-@v_args(inline=True)  # Affects the signatures of the methods
-class TransformToSQLAlchemyExpression(Transformer):
-    """Converts a field to a SQLAlchemy expression"""
-
-    def __init__(self, selectable, columns, drivername, forbid_aggregation=True):
-        self.text = None
-        self.selectable = selectable
-        self.columns = columns
-        self.last_datatype = None
-        # Convert all dates with this conversion
-        self.convert_dates_with = None
-        self.convert_datetimes_with = None
-        self.forbid_aggregation = forbid_aggregation
-        self.drivername = drivername
-
-    def _raise_error(self, message):
-        tree = None
-        tok = None
-        # Find the first token
-        while tree and tree.children:
-            tree = tree.children[0]
-            if isinstance(tree, Token):
-                tok = tree
-                break
-
-        if tok:
-            extra_context = self._get_context_for_token(tok)
-            message = f"{message}\n{extra_context}"
-        raise GrammarError(message)
-
-    def _get_context_for_token(self, tok, span=40):
-        pos = tok.start_pos
-        start = max(pos - span, 0)
-        end = pos + span
-        before = self.text[start:pos].rsplit("\n", 1)[-1]
-        after = self.text[pos:end].split("\n", 1)[0]
-        return before + after + "\n" + " " * len(before) + "^\n"
-
-    def col(self, v):
-        return v
-
-    def string(self, v):
-        return self.columns[v.data] if isinstance(v, Tree) else v
-
-    def string_cast(self, _, fld):
-        """Cast a field to a string"""
-        return cast(fld, String())
-
-    def string_substr(self, _, fld, *args):
-        """Substring a string. This can take one or two args."""
-        if self.drivername.startswith("mssql"):
-            if len(args) != 2:
-                raise GrammarError("mssql requires a starting number and a length")
-            return func.substring(fld, *args)
-
-        # Sqlite, postgres, bigquery, snowflake all accept an
-        # optional second argument for length
-        return func.substr(fld, *args)
-
-    def num(self, v):
-        return self.columns[v.data] if isinstance(v, Tree) else v
-
-    def int_cast(self, _, fld):
-        """Cast a field to a string"""
-        return cast(fld, Integer())
-
-    def coalesce(self, coalesce, left, right):
-        """Coalesce a number, string, date or datetime"""
-        return func.coalesce(left, right)
-
-    def boolean(self, v):
-        return self.columns[v.data] if isinstance(v, Tree) else v
-
-    def num_add(self, a, b):
-        """Add numbers or strings"""
-        return a + b
-
-    def string_add(self, a, b):
-        """Add numbers or strings"""
-        return a + b
-
-    def num_div(self, num, denom):
-        """SQL safe division"""
-        if not isinstance(denom, (int, float)):
-            return (
-                case([(denom == 0, None)], else_=num / cast(denom, Float))
-                if isinstance(num, (int, float))
-                else case(
-                    [(denom == 0, None)], else_=cast(num, Float) / cast(denom, Float)
-                )
-            )
-
-        if denom == 0:
-            raise GrammarError("When dividing, the denominator can not be zero")
-        elif denom == 1:
-            return num
-        elif isinstance(num, (int, float)):
-            return num / denom
-        else:
-            return cast(num, Float) / denom
-
-    def num_sub(self, a, b):
-        return a - b
-
-    def num_mul(self, a, b):
-        return a * b
-
-    def age_conv(self, _, fld):
-        """Convert a date to an age"""
-        if self.drivername == "bigquery":
-            return engine_support.bq_age(fld)
-        elif self.drivername == "sqlite":
-            raise GrammarError("Age is not supported on sqlite")
-        else:
-            # Postgres + redshift
-            return engine_support.postgres_age(fld)
-
-    # Dates and datetimes
-
-    def date(self, v):
-        if isinstance(v, Tree):
-            fld = self.columns[v.data]
-            if self.convert_dates_with:
-                converter = getattr(self, self.convert_dates_with, None)
-                if converter:
-                    fld = converter(None, fld)
-        else:
-            fld = v
-        return fld
-
-    def date_fn(self, _, y, m, d):
-        if self.drivername.startswith("mssql"):
-            return func.datefromparts(y, m, d)
-        else:
-            return func.date(y, m, d)
-
-    def datetime(self, v):
-        if isinstance(v, Tree):
-            fld = self.columns[v.data]
-            if self.convert_datetimes_with:
-                converter = getattr(self, self.convert_datetimes_with, None)
-                if converter:
-                    fld = converter(None, fld)
-        else:
-            fld = v
-        return fld
-
-    def datetime_end(self, v):
-        return self.columns[v.data] if isinstance(v, Tree) else v
-
-    def date_conv(self, _, datestr):
-        try:
-            dt = dateparser.parse(datestr)
-        except Exception as e:
-            raise e from e
-        if dt:
-            dt = dt.date()
-        else:
-            raise GrammarError(f"Can't convert '{datestr}' to a date.")
-        return dt
-
-    def datetime_conv(self, _, datestr):
-        dt = dateparser.parse(datestr)
-        if dt is None:
-            raise GrammarError(f"Can't convert '{datestr}' to a datetime.")
-        return dt
-
-    def day_conv(self, _, fld):
-        """Truncate to the day"""
-        if self.drivername == "bigquery":
-            return func.date_trunc(fld, text("day"))
-        elif self.drivername.startswith("mssql"):
-            return func.datefromparts(func.year(fld), func.month(fld), func.day(fld))
-        else:
-            # Postgres + redshift
-            return func.date_trunc("day", fld)
-
-    def week_conv(self, _, fld):
-        """Truncate to mondays"""
-        if self.drivername == "bigquery":
-            return func.date_trunc(fld, text("week(monday)"))
-        elif self.drivername.startswith("mssql"):
-            raise GrammarError("week is not supported on mssql")
-        else:
-            # Postgres + redshift
-            return func.date_trunc("week", fld)
-
-    def month_conv(self, _, fld):
-        if self.drivername == "bigquery":
-            return func.date_trunc(fld, text("month"))
-        elif self.drivername.startswith("mssql"):
-            return func.datefromparts(func.year(fld), func.month(fld), literal_1)
-        else:
-            # Postgres + redshift
-            return func.date_trunc("month", fld)
-
-    def quarter_conv(self, _, fld):
-        # Convert each date to the first day of each quarter
-        if self.drivername == "bigquery":
-            return func.date_trunc(fld, text("quarter"))
-        elif self.drivername.startswith("mssql"):
-            raise GrammarError("quarter is not supported on mssql")
-        else:
-            # Postgres + redshift
-            return func.date_trunc("quarter", fld)
-
-    def year_conv(self, _, fld):
-        if self.drivername == "bigquery":
-            return func.date_trunc(fld, text("year"))
-        elif self.drivername.startswith("mssql"):
-            return func.datefromparts(func.year(fld), literal_1, literal_1)
-        else:
-            # Postgres + redshift
-            return func.date_trunc("year", fld)
-
-    def dt_day_conv(self, _, fld):
-        """Truncate to day"""
-        if self.drivername == "bigquery":
-            # Developer's note: The timestamp_trunc function in sqlalchemy-bigquery is not
-            # type checked so the function return type is not annotated. This would
-            # generate a type error when running the query like:
-            # sqlalchemy.exc.DatabaseError: (google.cloud.bigquery.dbapi.exceptions.DatabaseError) 400
-            # No matching signature for operator BETWEEN for argument types: TIMESTAMP, DATETIME, DATETIME
-            #
-            # We'll explicitly convert which removes timezone info.
-            #
-            # I tried to create a custom function in engine_support but this kept
-            # creating a recursion error.
-            return func.datetime(func.timestamp_trunc(fld, text("day")))
-        elif self.drivername.startswith("mssql"):
-            return func.datetimefromparts(
-                func.year(fld),
-                func.month(fld),
-                func.day(fld),
-                literal_0,
-                literal_0,
-                literal_0,
-                literal_0,
-            )
-        else:
-            # Postgres + redshift
-            return func.date_trunc("day", fld)
-
-    def dt_week_conv(self, _, fld):
-        """Truncate to mondays"""
-        if self.drivername == "bigquery":
-            # See developer's note on dt_day_conv
-            return func.datetime(func.timestamp_trunc(fld, text("week(monday)")))
-        elif self.drivername.startswith("mssql"):
-            raise GrammarError("week is not supported on mssql")
-        else:
-            # Postgres + redshift
-            return func.date_trunc("week", fld)
-
-    def dt_month_conv(self, _, fld):
-        if self.drivername == "bigquery":
-            # See developer's note on dt_day_conv
-            return func.datetime(func.timestamp_trunc(fld, text("month")))
-        elif self.drivername.startswith("mssql"):
-            return func.datetimefromparts(
-                func.year(fld),
-                func.month(fld),
-                literal_1,
-                literal_0,
-                literal_0,
-                literal_0,
-                literal_0,
-            )
-        else:
-            # Postgres + redshift
-            return func.date_trunc("month", fld)
-
-    def dt_quarter_conv(self, _, fld):
-        if self.drivername == "bigquery":
-            # See developer's note on dt_day_conv
-            return func.datetime(func.timestamp_trunc(fld, text("quarter")))
-        elif self.drivername.startswith("mssql"):
-            raise GrammarError("quarter is not supported on mssql")
-        else:
-            # Postgres + redshift
-            return func.date_trunc("quarter", fld)
-
-    def dt_year_conv(self, _, fld):
-        if self.drivername == "bigquery":
-            # See developer's note on dt_day_conv
-            return func.datetime(func.datetime(func.timestamp_trunc(fld, text("year"))))
-        elif self.drivername.startswith("mssql"):
-            # SQL server can not support parameters in queries that are used for grouping
-            # https://github.com/mkleehammer/pyodbc/issues/479
-            # To avoid parameterization, we pass literals
-            return func.datetimefromparts(
-                func.year(fld),
-                literal_1,
-                literal_1,
-                literal_0,
-                literal_0,
-                literal_0,
-                literal_0,
-            )
-        else:
-            # Postgres + redshift
-            return func.date_trunc("year", fld)
-
-    def datetime_end_conv(self, _, datestr):
-        # Parse a datetime as the last moment of the given day
-        # if the date
-        dt = dateparser.parse(datestr)
-        if dt is None:
-            raise GrammarError(f"Can't convert '{datestr}' to a datetime.")
-        return convert_to_eod_datetime(dt)
-
-    def timedelta(self):
-        pass
-
-    # Booleans
-
-    def and_boolean(self, left_boolean, AND, right_boolean):
-        return and_(left_boolean, right_boolean)
-
-    def paren_boolean(self, paren_boolean):
-        return and_(paren_boolean)
-
-    def or_boolean(self, left_boolean, OR, right_boolean):
-        return or_(left_boolean, right_boolean)
-
-    def not_boolean(self, NOT, boolean_expr):
-        if boolean_expr in (True, False):
-            return not boolean_expr
-        else:
-            return not_(boolean_expr)
-
-    def consistent_array(self, *args):
-        """A comma separated, variable length array of all numbers
-        or all strings"""
-        return args
-
-    def vector_comparator(self, *args):
-        """Can be one token "IN" or two "NOT IN" """
-        return "in_" if len(args) == 1 else "notin_"
-
-    def comparator(self, comp):
-        """A comparator like =, !=, >, >="""
-        comparators = {
-            "=": "__eq__",
-            ">": "__gt__",
-            ">=": "__ge__",
-            "!=": "__ne__",
-            "<>": "__ne__",
-            "<": "__lt__",
-            "<=": "__le__",
-            "IS": "__eq__",
-            "ISNOT": "__ne__",
-        }
-        return comparators.get(str(comp).upper())
-
-    def null_comparator(self, *args):
-        comp = "="
-        if len(args) == 1:
-            comp = args[0]
-        elif len(args) == 2:
-            comp = "ISNOT"
-        return self.comparator(comp)
-
-    def between_expr(self, col, BETWEEN, left, AND, right):
-        return between(col, left, right)
-
-    def date_between_expr(self, col, BETWEEN, left, AND, right):
-        """Auto convert strings to dates."""
-        if isinstance(left, str):
-            left = self.date_conv(None, left)
-        if isinstance(right, str):
-            right = self.date_conv(None, right)
-        return self.between_expr(col, BETWEEN, left, AND, right)
-
-    def datetime_between_expr(self, col, BETWEEN, left, AND, right):
-        """Auto convert strings to datetimes."""
-        if isinstance(left, str):
-            left = self.datetime_conv(None, left)
-        if isinstance(right, str):
-            right = self.datetime_end_conv(None, right)
-        return self.between_expr(col, BETWEEN, left, AND, right)
-
-    def intelligent_date_expr(self, datecol, IS, offset, units):
-        start, end = calc_date_range(offset, units, date.today())
-        return between(datecol, start, end)
-
-    def intelligent_datetime_expr(self, datetimecol, IS, offset, units):
-        start, end = calc_date_range(offset, units, date.today())
-        start = convert_to_start_datetime(start)
-        end = convert_to_end_datetime(end)
-        return between(datetimecol, start, end)
-
-    def vector_expr(self, left, vector_comparator, num_or_str_array):
-        if hasattr(left, vector_comparator):
-            return getattr(left, vector_comparator)(num_or_str_array)
-        else:
-            self._raise_error("This value must be a column or column expression")
-
-    def bool_expr(self, left, comparator, right):
-        """A boolean expression like score > 20
-
-        If left is a primitive, swap the order:
-        20 > score => score < 20
-        """
-        # If the left is a primitive, try to swap the sides
-        if isinstance(left, (str, int, float, bool, date, datetime)):
-            swap_comp = {
-                "__gt__": "__lt__",
-                "__lt__": "__gt__",
-                "__ge__": "__le__",
-                "__le__": "__ge__",
-            }
-            comparator = swap_comp.get(comparator, comparator)
-            left, right = right, left
-
-        if right is None and comparator in ("__eq__", "__ne__"):
-            is_comp = {"__eq__": "is_", "__ne__": "isnot"}
-            comparator = is_comp.get(comparator, comparator)
-
-        return getattr(left, comparator)(right)
-
-    def date_bool_expr(self, left, comparator, right):
-        """If right is still a string, convert to a date."""
-        if isinstance(right, str):
-            right = self.date_conv(None, right)
-        return self.bool_expr(left, comparator, right)
-
-    def datetime_bool_expr(self, left, comparator, right):
-        """If right is still a string, convert to a date."""
-        if isinstance(right, str):
-            right = self.datetime_conv(None, right)
-        return self.bool_expr(left, comparator, right)
-
-    def str_like_expr(self, left, comparator, right):
-        """If right doesn't contain a wildcard, search for right
-        anywhere in the string."""
-        if "_" not in right and "%" not in right:
-            right = f"%{right}%"
-        if comparator.lower() == "like":
-            return left.like(right)
-        elif comparator.lower() == "ilike":
-            return left.ilike(right)
-        else:
-            raise Exception("Unknown comparator")
-
-    # Aggregations
-
-    def aggr(self, v):
-        return v
-
-    def date_aggr(self, v):
-        return v
-
-    def datetime_aggr(self, v):
-        return v
-
-    def string_aggr(self, v):
-        return v
-
-    def sum_aggr(self, _, fld):
-        """Sum up the things"""
-        return func.sum(fld)
-
-    def min_aggr(self, _, fld):
-        """Sum up the things"""
-        return func.min(fld)
-
-    def max_aggr(self, _, fld):
-        """Sum up the things"""
-        return func.max(fld)
-
-    def avg_aggr(self, _, fld):
-        """Sum up the things"""
-        return func.avg(fld)
-
-    def count_aggr(self, _, fld):
-        """Sum up the things"""
-        if getattr(fld, "data", None) == "star":
-            return func.count()
-        else:
-            return func.count(fld)
-
-    def percentile_aggr(self, percentile, fld):
-        """Sum up the things"""
-        percentile_val = int(percentile[len("percentile") :])
-        if percentile_val not in (1, 5, 10, 25, 50, 75, 90, 95, 99):
-            raise GrammarError(
-                f"percentile values of {percentile_val} is not supported."
-            )
-        if self.drivername == "bigquery":
-            percentile_fn = getattr(engine_support, f"bq_percentile{percentile_val}")
-            return percentile_fn(fld)
-        elif self.drivername == "sqlite":
-            raise GrammarError("Percentile is not supported on sqlite")
-        else:
-            # Postgres + redshift
-            return func.percentile_cont(percentile_val / 100.0).within_group(fld)
-
-    def count_distinct_aggr(self, _, fld):
-        """Sum up the things"""
-        return func.count(distinct(fld))
-
-    # If functions
-
-    def if_statement(self, IF, *args):
-        args = list(args)
-
-        # If there's an odd number of args, pop the last one to use as the else
-        if len(args) % 2:
-            else_expr = args.pop()
-        else:
-            else_expr = None
-
-        # collect the other args into pairs
-        # ['a','b','c','d'] --> [('a',b'), ('c','d')]
-        pairs = zip(args[::2], args[1::2])
-        return case(pairs, else_=else_expr)
-
-    # Constants
-
-    def ESCAPED_STRING(self, v):
-        v = str(v)
-        if v.startswith('"') and v.endswith('"'):
-            return v[1:-1]
-        return v
-
-    def NUMBER(self, v):
-        try:
-            n = int(v)
-        except ValueError:
-            n = float(v)
-        return n
-
-    def TRUE(self, v):
-        return True
-
-    def FALSE(self, v):
-        return False
-
-    def NULL(self, v):
-        return None
-
-
-LARK_CACHE = {}
-
-
-class SQLAlchemyBuilder(object):
-    @classmethod
-    def get_builder(cls, selectable, *, cache=None):
-        return cls(selectable, cache=cache)
-
-    @classmethod
-    def clear_builder_cache(cls):
-        LARK_CACHE.clear()
-
-    def __init__(self, selectable, *, cache=None):
-        """Parse a recipe field by building a custom grammar that
-        uses the colums in a selectable.
-
-        Args:
-            selectable (Table): A SQLAlchemy selectable
-        """
-        self.selectable = selectable
-        # Database driver
-        try:
-            self.drivername = selectable.metadata.bind.url.drivername
-        except Exception:
-            self.drivername = "unknown"
-
-        self.cache = cache
-        self.columns = make_columns_for_selectable(selectable)
-        self.grammar = make_grammar(self.columns)
-        grammar_hash = mkkey("grammar", self.grammar)
-        # Developer Note: cache key
-        # This cache key is used for both LARK_CACHE as well as the SQLAlchemy
-        # expressions that we use in `parse` below. This key must change any time the
-        # table columns change, which it *should* do because the grammar contains
-        # information about all types and columns in the table. If that ever changes
-        # (e.g. we switch to separate steps for parsing and type/variable validation),
-        # we'll need to calculate the expression hash key separately.
-        self.cache_key = f"recipe-expr:{grammar_hash}"
-        try:
-            self.cached_trees = (
-                self.cache.get(self.cache_key, {}) if self.cache is not None else None
-            )
-        except Exception:
-            SLOG.exception("ingredient-cache-error")
-            self.cached_trees = None
-        if self.cache_key in LARK_CACHE:
-            self.parser = LARK_CACHE[self.cache_key]
-        else:
-            # Constructing this Lark parser can take a significant amount of time (like,
-            # nearly 1 second for some large tables), which is why we cache parsers in
-            # LARK_CACHE. Unfortunately, by using an in-process cache, we are still
-            # computing this redundantly quite often, because Juicebox processes get
-            # cycled often and there are many workers in a given deployment.
-            #
-            # Lark supports serializing parsers to speed up loading, but unfortunately
-            # it only supports this for LALR parsers, and we are using Earley.
-            self.parser = Lark(
-                self.grammar,
-                parser="earley",
-                ambiguity="resolve",
-                start="col",
-                propagate_positions=True,
-                # predict_all=True,
-            )
-            LARK_CACHE[self.cache_key] = self.parser
-
-        self.transformer = TransformToSQLAlchemyExpression(
-            self.selectable, self.columns, self.drivername
-        )
-
-        # The data type of the last parsed expression
-        self.last_datatype = None
-
-    def parse(
-        self,
-        text,
-        forbid_aggregation=False,
-        enforce_aggregation=False,
-        debug=False,
-        convert_dates_with=None,
-        convert_datetimes_with=None,
-    ):
-        """Return a parse tree for text
-
-        Args:
-            text (str): A field expression
-            forbid_aggregation (bool, optional):
-              The expression may not contain aggregations. Defaults to False.
-            enforce_aggregation (bool, optional):
-              Wrap the expression in an aggregation if one is not provided. Defaults to False.
-            debug (bool, optional): Show some debug info. Defaults to False.
-            convert_dates_with (str, optional): A converter to use for date fields
-            convert_datetimes_with (str, optional): A converter to use for datetime fields
-
-        Raises:
-            GrammarError: A description of any errors and where they occur
-
-        Returns:
-            A tuple of
-                ColumnElement: A SQLALchemy expression
-                DataType: The datatype of the expression (bool, date, datetime, num, str)
-        """
-        key = mkkey(
-            "parsed-ingredient",
-            text,
-            forbid_aggregation,
-            enforce_aggregation,
-            convert_dates_with,
-            convert_datetimes_with,
-        )
-        cache_result = (
-            self.cached_trees.get(key) if self.cached_trees is not None else None
-        )
-
-        extra_args = (
-            key,
-            enforce_aggregation,
-            debug,
-            convert_dates_with,
-            convert_datetimes_with,
-        )
-
-        if cache_result is None:
-            (tree, validator) = self._parse(text, forbid_aggregation)
-            return self.tree_to_expression(tree, validator, *extra_args)
-        else:
-            (tree, validator) = cache_result
-            try:
-                return self.tree_to_expression(tree, validator, *extra_args)
-            except Exception:
-                SLOG.exception("cached-tree-to-validator-error")
-                # If we get ANY error while dealing with the cached ingredient data, we
-                # should just retry everything without using the cache. There could be
-                # any number of things wrong with the cache, e.g. if it was produced on
-                # an older version of Recipe (there are a lot of internal implementation
-                # details encoded into the cached data).
-                del self.cached_trees[key]
-                (tree, validator) = self._parse(text, forbid_aggregation)
-                return self.tree_to_expression(tree, validator, *extra_args)
-
-    def _parse(self, text, forbid_aggregation):
-        tree = self.parser.parse(text, start="col")
-        validator = SQLALchemyValidator(text, forbid_aggregation, self.drivername)
-        validator.visit(tree)
-        return (tree, validator)
-
-    def tree_to_expression(
-        self,
-        tree,
-        validator,
-        key,
-        enforce_aggregation,
-        debug,
-        convert_dates_with,
-        convert_datetimes_with,
-    ):
-        self.last_datatype = validator.last_datatype
-        if validator.errors:
-            if debug:
-                print("".join(validator.errors))
-                print("Tree:\n" + tree.pretty())
-            raise GrammarError("".join(validator.errors))
-
-        if debug:
-            print("Tree:\n" + tree.pretty())
-        self.transformer.text = text
-        self.transformer.convert_dates_with = convert_dates_with
-        self.transformer.convert_datetimes_with = convert_datetimes_with
-        expr = self.transformer.transform(tree)
-
-        # Expressions that return literal values can't be labeled
-        # Possibly we could wrap them in text() but this may be unsafe
-        # instead we will disallow them.
-        if isinstance(expr, (str, float, int, date, datetime)):
-            raise GrammarError("Must return an expression, not a constant value")
-
-        if (
-            enforce_aggregation
-            and not validator.found_aggregation
-            and self.last_datatype == "num"
-        ):
-            result = (func.sum(expr), self.last_datatype)
-        else:
-            result = (expr, self.last_datatype)
-
-        if self.cached_trees is not None:
-            self.cached_trees[key] = (tree, validator)
-        return result
-
-    def save_cache(self):
-        # see "Developer Note: cache key" for info about cache keys.
-        try:
-            self.cache.set(self.cache_key, self.cached_trees)
-        except Exception:
-            SLOG.exception("shelf-save-cache-error")
-
-
-def mkkey(prefix, *parts):
-    hash = hashlib.sha1()
-    for k in parts:
-        hash.update(str(k).encode("utf-8"))
-    return f"{prefix}-{hash.hexdigest()}"
