+++ recipe/schemas/lark_grammar.py
@@ -1,2 +1,3 @@
 # For compatibility
 from .expression_grammar import *
+from .builders import SQLAlchemyBuilder
